+++ Application/Transactions/ITransactionService.cs
@@ -4,7 +4,7 @@ public interface ITransactionService
 {
     Task<bool> AddDepositAsync(OperationDto operation);
     Task<bool> WithdrowAsync(OperationDto operation);
-    Task<float> GetBalanceAsync(int accountId);
+    Task<double> GetBalanceAsync(int accountId);
     Task<PagedList<TransactionDto>> GetTransactionsAsync(TransactionParams transactionParams);
     Task<TransactionDto> GetTransactionAsync(int id);
 }
